pyproject.toml
[project]
name = "trading-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "huggingface-hub>=1.9.2",
    "jax>=0.9.2",
    "jaxlib>=0.9.2",
    "matplotlib>=3.10.8",
    "mplfinance>=0.12.10b0",
    "numpy>=2.4.4",
    "pandas==2.2.3",
    "pandas-datareader>=0.10.0",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "rich>=14.3.3",
    "schedule>=1.2.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "setuptools>=82.0.1",
    "shap>=0.51.0",
    "timesfm>=1.3.0",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "yfinance>=1.2.1",
]
